requirements.txt
flask
python-dotenv
tqdm
pandas
numpy
torch
sentence-transformers
openai-whisper
google-generativeai
flask_cors
flask_sqlalchemy
dotenv
psycopg2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2